feat: 결제 로직 리팩토링
