+++ src/main/java/com/myce/payment/service/portone/PortOneApiService.java
@@ -1,0 +1,12 @@
+package com.myce.payment.service.portone;
+
+import com.myce.payment.dto.PaymentRefundRequest;
+import java.util.Map;
+
+public interface PortOneApiService {
+    String getAccessToken();
+
+    Map<String, Object> getPaymentInfo(String impUid, String accessToken);
+
+    Map<String, Object> requestRefundToPortOne(PaymentRefundRequest request, Integer originalPaidAmount, String accessToken);
+}
