+++ src/main/java/com/myce/payment/service/refund/PaymentRefundService.java
@@ -1,0 +1,8 @@
+package com.myce.payment.service.refund;
+
+import com.myce.payment.dto.PaymentRefundRequest;
+import java.util.Map;
+
+public interface PaymentRefundService {
+    Map<String, Object> refundPayment(PaymentRefundRequest request);
+}
