+++ src/main/java/com/myce/payment/service/webhook/PaymentWebhookService.java
@@ -1,0 +1,7 @@
+package com.myce.payment.service.webhook;
+
+import com.myce.payment.dto.PortOneWebhookRequest;
+
+public interface PaymentWebhookService {
+    void processWebhook(PortOneWebhookRequest request);
+}
